componentized images, working on applying vue defineprops
