+++ 05-Vue_Expanded/assignment/fishcamp/package.json
@@ -1,6 +1,6 @@
 {
   "name": "fishcamp",
-  "version": "0.0.0",
+  "version": "0.0.1",
   "private": true,
   "type": "module",
   "scripts": {
@@ -25,6 +25,6 @@
     "eslint": "^8.57.0",
     "eslint-plugin-vue": "^9.23.0",
     "prettier": "^3.2.5",
-    "vite": "^5.3.1"
+    "vite": "^5.4.6"
   }
 }
