+++ 05-Vue_Expanded/assignment/fishcamp/src/App.vue
@@ -5,6 +5,8 @@ import Accordion from './components/Accordion.vue'
 import Contact from './components/Contact.vue'
 import Gallery from './components/Gallery.vue'
 import NotFound from './components/NotFound.vue'
+import Header from './components/Header.vue'
+import Footer from './components/Footer.vue'
 
 //use standard SPA routing from docs or assignment video, map path routes, then rendered values in object called routes
 const routes = {
@@ -30,59 +32,14 @@ const currentView = computed(() => {
 </script>
 
 <template>
-  <header class="row bg-primary align-items-center p-0 m-0">
-    <div class="col-lg-3 col-xs-10 p-2 text-center">
-      <img
-        class="img-fluid"
-        src="./assets/image.png"
-        alt="company logo, browser not compatible"
-        width="200"
-        height="200"
-      />
-    </div>
-    <div class="col-lg-5 col-xs-10 p-2 text-center">
-      <h2 class="">FishCamp, LLC Guided Fishing</h2>
-      <h4 class="">
-        Charters and guided and backpack fishing sustainably in WA, OR, ID mountain lakes and
-        streams
-      </h4>
-    </div>
-    <div class="col-lg-4 col-xs-10 p-2 text-center">
-      <h5 class="">12345 McRoad Street, Black Diamond, WA 98010</h5>
-    </div>
-  </header>
-  <div class="row bg-warning p-0 m-0">
-    <nav class="navbar navbar-dark btn-group-horizontal justify-content-around">
-      <a href="#/"><button class="btn btn-outline-primary m-3">Landing Page</button></a>
+  <Header></Header>
 
-      <a href="#/Gallery"><button class="btn btn-outline-primary m-3">Trip Gallery</button></a>
-
-      <a href="#/Accordion"><button class="btn btn-outline-primary m-3">FAQ Page</button></a>
-
-      <a href="#/Contact"><button class="btn btn-outline-primary m-3">Contact Us</button></a>
-    </nav>
-  </div>
   <main class="row bg-white p-0 m-0">
     <!--main content element that is dependent on the logic in script above-->
     <component :is="currentView" />
   </main>
 
-  <footer class="row text-light bg-dark text-center p-0 m-0">
-    <h4 class="p-4 m-3">FishCamp LLC - Sustaining Natural Memories</h4>
-    <p class="lead p-4">
-      Charters and guided and backpack fishing sustainably in WA, OR, ID mountain lakes and streams
-    </p>
-    <nav class="navbar navbar-dark btn-group-vertical justify-content-around">
-      <a href="./booking.html"><button class="btn btn-outline-primary m-3">Book A Trip</button></a>
-
-      <a href="./gallery.html"><button class="btn btn-outline-primary m-3">Trip Gallery</button></a>
-
-      <a href="./faq.html"><button class="btn btn-outline-primary m-3">FAQ Page</button></a>
-
-      <a href="./contact.html"><button class="btn btn-outline-primary m-3">Contact Us</button></a>
-    </nav>
-    <p class="p-4">Created 2024 Sites-To-Go LLC Seattle</p>
-  </footer>
+  <Footer></Footer>
 </template>
 
 <style scoped></style>
