+++ 05-Vue_Expanded/assignment/fishcamp/src/components/Footer.vue
@@ -1,0 +1,26 @@
+<script setup></script>
+
+<template>
+  <div class="row text-light bg-dark text-center p-0 m-0">
+    <h4 class="p-4 m-3">FishCamp LLC - Sustaining Natural Memories</h4>
+    <p class="lead p-4">
+      Charters and guided and backpack fishing sustainably in WA, OR, ID mountain lakes and streams
+    </p>
+    <nav class="navbar navbar-dark btn-group-vertical justify-content-around">
+      <a href="#/"><button class="btn btn-outline-primary m-3">Landing Page</button></a>
+
+      <a href="#/Gallery"><button class="btn btn-outline-primary m-3">Trip Gallery</button></a>
+
+      <a href="#/Accordion"><button class="btn btn-outline-primary m-3">FAQ Page</button></a>
+
+      <a href="#/Contact"><button class="btn btn-outline-primary m-3">Contact Us</button></a>
+    </nav>
+    <p class="p-4">Created 2024 Sites-To-Go LLC Seattle</p>
+  </div>
+</template>
+
+<script>
+export default {}
+</script>
+
+<style scoped></style>
